+++ src/compas_fab/backends/vrep/backend_features/__init__.py
@@ -1,4 +1,5 @@
 from .vrep_add_attached_collision_mesh import VrepAddAttachedCollisionMesh
+from .vrep_add_collision_mesh import VrepAddCollisionMesh
 from .vrep_forward_kinematics import VrepForwardKinematics
 from .vrep_plan_motion import VrepPlanMotion
 from .vrep_inverse_kinematics import VrepInverseKinematics
@@ -7,5 +8,6 @@ __all__ = [
     'VrepForwardKinematics',
     'VrepInverseKinematics',
     'VrepPlanMotion',
-    'VrepAddAttachedCollisionMesh'
+    'VrepAddAttachedCollisionMesh',
+    'VrepAddCollisionMesh',
 ]
