+++ src/compas_fab/backends/vrep/planner_backend_vrep.py
@@ -4,6 +4,7 @@ from __future__ import print_function
 
 from compas_fab.backends.client import PlannerInterface
 from compas_fab.backends.vrep.backend_features.vrep_add_attached_collision_mesh import VrepAddAttachedCollisionMesh
+from compas_fab.backends.vrep.backend_features.vrep_add_collision_mesh import VrepAddCollisionMesh
 from compas_fab.backends.vrep.backend_features.vrep_forward_kinematics import VrepForwardKinematics
 from compas_fab.backends.vrep.backend_features.vrep_inverse_kinematics import VrepInverseKinematics
 from compas_fab.backends.vrep.backend_features.vrep_plan_motion import VrepPlanMotion
@@ -18,3 +19,4 @@ class VrepPlanner(PlannerInterface):
         self.plan_motion_to_config = VrepPlanMotion(self.client).plan_motion_to_config
         self.add_attached_collision_mesh = VrepAddAttachedCollisionMesh(self.client)
         self.pick_building_member = VrepAddAttachedCollisionMesh(self.client).pick_building_member
+        self.add_collision_mesh = VrepAddCollisionMesh(self.client)
